Split glob parsing out of the PathGlobs struct to make it more analogous to the Python-side object.
[ci skip-jvm-tests]
